[Fix] Stay에 guestFavourite 수정, PlaceType, PropertyType 한글로 변경 - guestFavourite은 나중에 필요할 수도 있어서 일단 주석 처리했습니다
